[DART] merged both strategy implementations.
