+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/bank_strategy.dart
@@ -1,4 +1,4 @@
-import 'package:dart_lang/b_behavioral/strategy/payment_strategy.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/other_strategy_implementations/payment_strategy.dart';
 
 class BankStrategy implements PaymentStrategy {
   @override
