+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/exercise/notification_sending_service.dart
@@ -1,0 +1,5 @@
+import 'package:dart_lang/b_behavioral/b_strategy/other_strategy_implementations/exercise/user.dart';
+
+abstract class NotificationSendingService {
+  void sendNotification(String message, User user);
+}
