+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/exercise/notification_services.dart
@@ -1,5 +1,5 @@
-import 'package:dart_lang/b_behavioral/strategy/exercise/notification_sending_service.dart';
-import 'package:dart_lang/b_behavioral/strategy/exercise/user.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/other_strategy_implementations/exercise/notification_sending_service.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/other_strategy_implementations/exercise/user.dart';
 
 class NotifcationServices {
   final NotificationSendingService notificationSendingService;
