+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/gold_pricing_strategy.dart
@@ -1,4 +1,4 @@
-import 'package:dart_lang/b_behavioral/strategy/pricing_strategy.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/other_strategy_implementations/pricing_strategy.dart';
 
 class GoldPricingStartge implements PricingStartgy {
   @override
