+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/payment_strategy.dart
