+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/pricing_strategy.dart
