+++ dart_lang/lib/b_behavioral/b_strategy/other_strategy_implementations/visa_card_startegy.dart
@@ -1,4 +1,4 @@
-import 'package:dart_lang/b_behavioral/strategy/payment_strategy.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/other_strategy_implementations/payment_strategy.dart';
 
 class VisaCardStratge implements PaymentStrategy {
   @override
