--- dart_lang/lib/b_behavioral/strategy/exercise/notification_sending_service.dart
@@ -1,5 +1,0 @@
-import 'package:dart_lang/b_behavioral/strategy/exercise/user.dart';
-
-abstract class NotificationSendingService {
-  void sendNotification(String message, User user);
-}
