--- dart_lang/lib/b_behavioral/strategy/exercise/send_mail.dart
@@ -1,9 +1,0 @@
-import 'package:dart_lang/b_behavioral/strategy/exercise/notification_sending_service.dart';
-import 'package:dart_lang/b_behavioral/strategy/exercise/user.dart';
-
-class SendingMail implements NotificationSendingService {
-  @override
-  void sendNotification(String message, User user) {
-    print('sending $message to ${user.email}');
-  }
-}
